moved jdbc service into correct folder
removed unused h3 jdbc beans from other services
