+++ src/main/java/com/dedicatedcode/reitti/service/h3/AreaBoundaryLookupBase.java
@@ -2,6 +2,7 @@ package com.dedicatedcode.reitti.service.h3;
 
 import com.dedicatedcode.reitti.dto.area.AreaDescription;
 import com.dedicatedcode.reitti.dto.area.AreaBounds;
+import com.dedicatedcode.reitti.repository.h3.AreaJdbcService;
 
 import java.util.Comparator;
 import java.util.List;
