+++ src/main/java/com/dedicatedcode/reitti/service/h3/H3AreaMappingService.java
@@ -1,6 +1,7 @@
 package com.dedicatedcode.reitti.service.h3;
 
 import com.dedicatedcode.reitti.config.H3MappingConfig;
+import com.dedicatedcode.reitti.repository.h3.AreaJdbcService;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.boot.autoconfigure.condition.ConditionalOnProperty;
